django/requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
black==20.8b1
click==7.1.2
Django==3.1.7
django-cors-headers==3.7.0
django-graphql-jwt==0.3.1
django-js-asset==1.2.2
django-mptt==0.12.0
djangorestframework==3.12.2
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
isort==5.7.0
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.1.2
promise==2.3
PyJWT==1.7.1
pytz==2021.1
regex==2020.11.13
Rx==1.6.1
singledispatch==3.6.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
